Don't exit if there are no repo changes
Currently, the Jenkins build is triggered when there is an SCM change.
So, if there is an SCM change, there wil be a gitlog. Hence, we don't
really need to check for the gitlog file's existence.

Now, there will be scenarios, where we want to trigger an explicit
build. Re-builds ? Changes in debian/ repo ?
Hence, again, we shouldn't rely on gitlog file.