Determine and build package
When run under Jenkins, we get the BUILD_NUMBER environment variable.
That is what is already build used, when formatting the changelog.

In this commit, we are re-using the same environment variable. If the
variable is unset, it is safe to assume that we are being built on a
developer box. In that case, we set the version accordingly.

Repetitive builds on the developer box are guaranteed to of >N version